requirements.txt
pydantic==1.10.2
email-validator==1.2.1
fastapi==0.68.1
PyJWT==2.4.0
python-decouple==3.6
uvicorn==0.18.3
passlib~=1.7.4